Time: 864 ms (25.97%), Space: 41.6 MB (49.35%) - LeetHub

+++ 1680-concatenation-of-consecutive-binary-numbers/1680-concatenation-of-consecutive-binary-numbers.java
@@ -2,7 +2,7 @@ class Solution {
     public int concatenatedBinary(int n) {
         int mod = 1000000007;
         
-        long ans = 0;
+        int ans = 0;
         for(int i=1;i<=n;i++){
             int x = i;
             while(x>0){
@@ -12,6 +12,6 @@ class Solution {
             
             ans = (ans + i)%mod;
         }
-        return (int)ans;
+        return ans;
     }
 }
